smoke_tests/requirements.txt
# qTest API Smoke Tests - Python Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
